pyproject.toml
[tool.poetry]
name = "bookjibe"
version = "0.1.0"
description = ""
authors = ["Jean-Baptiste Poullet"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.12.0"
langchain = "^0.1.8"
langchainhub = "^0.1.14"
langchain-openai = "^0.0.6"
python-dotenv = "^1.0.1"
dash = "^2.15.0"
pandas = "^2.2.0"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
